+++ docs/changelog.rst
@@ -78,6 +78,12 @@ Any breaking changes which require intervention will be mentioned here.
   generation.
 - The :meth:`.ReceiptPDF.save_pdf` method now optionally takes an instance of
   the above :class:`~.PdfBuilder`.
+- **BREAKING**: The ``raise_`` flag for :meth:`~.Receipt.validate` now defaults
+  to ``True``. I.e: failure to validate a ``Receipt`` will raise a
+  :class:`~.ValidationError` exception by default. This is the pythonic
+  approach and should prevent accidental situations where errors go unnoticed
+  and are not handled properly. The ``raise_`` flag is still deprecated and
+  pending removal.
 - Type hints have been added everywhere that is feasible.
 - Add a new helper helper method :meth:`.Receipt.approximate_date`. It is
   intended to be used to automatically approximate dates on systems which